requirements.txt
seaborn
timm
pretrainedmodels
mlflow
jupyterlab
scikit-multilearn
absl-py==0.11.0
aiohttp>=3.7.4
appdirs==1.4.4
appnope==0.1.0
astor==0.7.1
astroid==2.3.3
astunparse==1.6.3
async-timeout>=4
attrs==20.3.0
backcall==0.1.0
bleach>=3.3.0
cachetools==4.2.0
certifi==2020.4.5.1
cffi==1.14.4
chardet==3.0.4
click==7.1.2
colorgram.py==1.2.0
cryptography>=3.3.2
cycler==0.10.0
dataclasses==0.6
decorator==4.4.2
docutils==0.16
envyaml==0.2060
Flask==1.1.2
fsspec==0.8.5
future==0.18.2
gast==0.3.3
gcsfs==0.7.1
google-api-core==1.24.1
google-api-python-client==1.12.8
google-auth==1.24.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
google-cloud-core==1.5.0
google-cloud-storage==1.35.0
google-crc32c==1.1.0
google-pasta==0.2.0
google-resumable-media==1.2.0
googleapis-common-protos==1.52.0
grpcio
h5py==2.10.0
httplib2>=0.19.0
idna==2.9
imageio==2.9.0
imbalanced-learn==0.8.0
imblearn==0.0
imgaug==0.4.0
importlib-metadata==1.6.0
ipython-genutils==0.2.0
isort==4.3.21
itsdangerous==1.1.0
jedi==0.17.0
jeepney==0.6.0
jinja2>=2.11.3
joblib==0.14.1
Keras==2.4.3
Keras-Preprocessing==1.1.2
keyring==21.2.0
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
lightgbm==2.3.1
loguru==0.5.2
mahotas==1.4.11
Markdown==3.0.1
MarkupSafe>=2.0
marshmallow==3.7.1
matplotlib==3.2.1
mccabe==0.6.1
multidict==5.1.0
networkx==2.5
numpy>= 1.19.5
oauth2client==4.1.3
oauthlib==3.1.0
opencv-python==4.5.1.48
opencv-python-headless==4.2.0.34
opt-einsum==3.3.0
packaging==20.7
pandas==0.24.2
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow>=8.4.0
pkginfo==1.5.0.1
protobuf==3.15.0
psycopg2-binary==2.8.6
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments>=2.7.4
pyheif==0.5.1
pylint==2.4.4
pyparsing==2.4.7
python-dateutil==2.8.1
python-resize-image==1.1.19
pytz==2019.3
PyWavelets==1.1.1
PyYAML>=5.4
readme-renderer==25.0
regex==2020.4.4
requests>2.24.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rsa>=4.7
scikit-image==0.17.2
scikit-learn==0.24.1
scipy==1.4.1
SecretStorage==3.3.0
Shapely==1.7.1
six==1.15.0
termcolor==1.1.0
threadpoolctl==2.1.0
tifffile==2020.9.3
tqdm==4.45.0
traitlets
twine==3.1.1
typed-ast==1.4.0
typing-extensions==3.7.4.3
tzlocal==2.0.0
uritemplate==3.0.1
urllib3>=1.26.5
wcwidth==0.1.9
webargs==6.1.0
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.11.1
xgboost==1.2.1
yarl==1.6.3
zipp==3.1.0



